resolved issue with xterm and gray scales (possibly fixes other terms too)
gnome-terminal and konsole and others all use the setab and setaf
format strings to output gray scales but xterm uses setb and setf
to do this (xterm is probably correct).  to handle this in both
cases I have created a hard coded format string for gray scales
instead of using the usual ti_setaf() and ti_setab() macros.

added code to dump entire contents of the escape sequence buffer
of every frame of the display which is how i discovered the cause
of the above problem.  p.s. do not think xterm is doing this wrong