fixed some bugs related to double wide characters co-existing with single wide
when a double wide character such as chinese is written to the display it
takes up two character cells [XX].  it must therefore also take up two
character cells in our window/screen buffers.  To prevent the second cell
being displayed we set its codepoint to 0xDEADCODE.  The attributes of this
ccell will be the same as its parent character to the left (important later)

if we write a double wide character to a window but then overlay the 0xDEAD
cell with a single wide char so [XX] becones [Xx]  the cell following the double
wide character is no longer a DEADCODE and we must FORCE an update of that
character.  The double wide char WILl be written to the display and will
obliterate the single wide char from the display.  Even if the single wide char
was not modified it must now still be updated.

also.  If we overwrite the live cell of a double wide character with a single
wide character the following 0xDEAD cell will need to be changed to a space
so it wont be skipped and leave glitches all over the display.