building with gcc now results in a working library
when parsing the json file, after completely populating the screen
structure with all of the child structures specified in the json
the code would attempt to populate the screens parent with the
screeb structure.  This is kind of silly because the screen is the
root of the user interface and has no parent.

the real question here is how in the name of keyush the stunt dog
did this ever work when compiled with clang? (p.s. it worked fine!)