Totally refactored the demo code in preparation for proper demos.
Updated the readme.

Added ability for user applications to stuff any key into the keyboard
buffer as if a key had been pressed. Acutually this was already there
but not visible to user apps.

Fixed a bug in the json code allowing for user applications to redefine
their interface depending on context. I.E.  Scan json, construct UI.
then destroy current UI and scan a different json to construct a new one

Also fixed a bug related to specifying window widths and heights as
percentages of the screen size. Now accounts for when said window also
has a border (i think this is right now?)

Fixed bug related to calculating character column width on the console.
this still needs some more work but is now handled more accurately i
think.

Fixed a small bug in win_printf().  This code has not yet been fully
exercised so i expect to see more.

Added a spash screen to the demo code with instructions too! (yay!)