FINALLY Added widgets to user interface (still work in progress)
A lot of changes here and not all are strictly related to the
addition of widgets to the user interface but most were an
absolute requirement for my sanity.  trust me kthxbai!

made as many of the structures and enumerations in the sources
attribute packed.  the compiler complains on some of them though

added make install for of all header files with a uC_ prefix.

removed the uC_ prefix on any header file that should not be
referenced by the user.  these files are not installed by make
install

modified the generic names of all the color enumerations

made the attribute structure significantly more manageable and
less confusificating and allowing end users to more easily define
attributes

added ability to draw a box within a window using the same border
characters as apply to windows.  these are not sub windows

added 99.999% of the missing terminfo format string names and
offsets but many of these will be deleted again because they are
related to ancient printer control. these are not accessable to
the end user without a API uC_xxxx() wrapper which not all have

made the terminfo parsing variables more 'structured'

implemented a mechanism where you can now embed your user interfaces
json file into your binary.   added rant to src/uCurses.c related
to this (C is a fucking horrible language)